Move rule type engine code to more descriptive files
