+++ internal/engine/rule_type_engine.go
@@ -21,9 +21,7 @@ import (
 	"strings"
 
 	"github.com/rs/zerolog/log"
-	"github.com/xeipuuv/gojsonschema"
 	"google.golang.org/protobuf/encoding/protojson"
-	"google.golang.org/protobuf/types/known/structpb"
 
 	"github.com/stacklok/minder/internal/db"
 	"github.com/stacklok/minder/internal/engine/actions"
@@ -58,102 +56,9 @@ func (r *RuleMeta) String() string {
 	return fmt.Sprintf("%s/org/%s/%s", r.Provider, *r.Organization, r.Name)
 }
 
-// RuleValidator validates a rule against a schema
-type RuleValidator struct {
-	ruleTypeName string
-
-	// schema is the schema that this rule type must conform to
-	schema *gojsonschema.Schema
-	// paramSchema is the schema that the parameters for this rule type must conform to
-	paramSchema *gojsonschema.Schema
-}
-
-// NewRuleValidator creates a new rule validator
-func NewRuleValidator(rt *minderv1.RuleType) (*RuleValidator, error) {
-	// Load schemas
-	schemaLoader := gojsonschema.NewGoLoader(rt.Def.RuleSchema)
-	schema, err := gojsonschema.NewSchema(schemaLoader)
-	if err != nil {
-		return nil, fmt.Errorf("cannot create json schema: %w", err)
-	}
-
-	var paramSchema *gojsonschema.Schema
-	if rt.Def.ParamSchema != nil {
-		paramSchemaLoader := gojsonschema.NewGoLoader(rt.Def.ParamSchema)
-		paramSchema, err = gojsonschema.NewSchema(paramSchemaLoader)
-		if err != nil {
-			return nil, fmt.Errorf("cannot create json schema for params: %w", err)
-		}
-	}
-
-	return &RuleValidator{
-		ruleTypeName: rt.Name,
-		schema:       schema,
-		paramSchema:  paramSchema,
-	}, nil
-}
-
-// ValidateRuleDefAgainstSchema validates the given contextual profile against the
-// schema for this rule type
-func (r *RuleValidator) ValidateRuleDefAgainstSchema(contextualProfile map[string]any) error {
-	if err := validateAgainstSchema(r.schema, contextualProfile); err != nil {
-		return &RuleValidationError{
-			RuleType: r.ruleTypeName,
-			Err:      err.Error(),
-		}
-	}
-
-	return nil
-}
-
-// ValidateParamsAgainstSchema validates the given parameters against the
-// schema for this rule type
-func (r *RuleValidator) ValidateParamsAgainstSchema(params *structpb.Struct) error {
-	if r.paramSchema == nil {
-		return nil
-	}
-
-	if params == nil {
-		return &RuleValidationError{
-			RuleType: r.ruleTypeName,
-			Err:      "params cannot be nil",
-		}
-	}
-
-	if err := validateAgainstSchema(r.paramSchema, params.AsMap()); err != nil {
-		return &RuleValidationError{
-			RuleType: r.ruleTypeName,
-			Err:      err.Error(),
-		}
-	}
-
-	return nil
-}
-
-func validateAgainstSchema(schema *gojsonschema.Schema, obj map[string]any) error {
-	documentLoader := gojsonschema.NewGoLoader(obj)
-	result, err := schema.Validate(documentLoader)
-	if err != nil {
-		return fmt.Errorf("cannot validate json schema: %s", err)
-	}
-
-	if !result.Valid() {
-		return buildValidationError(result.Errors())
-	}
-
-	return nil
-}
-
-func buildValidationError(errs []gojsonschema.ResultError) error {
-	problems := make([]string, 0, len(errs))
-	for _, desc := range errs {
-		problems = append(problems, desc.String())
-	}
-
-	return fmt.Errorf("invalid json schema: %s", strings.TrimSpace(strings.Join(problems, "\n")))
-}
-
-// RuleTypeEngine is the engine for a rule type
+// RuleTypeEngine is the engine for a rule type. It builds the multiple
+// sections of the rule type and instantiates the needed drivers for
+// them.
 type RuleTypeEngine struct {
 	Meta RuleMeta
